aw_scraper-main/requirements.txt
Flask == 3.0.3
Scrapy == 2.11.2
scrapyrt == 0.16.0
yfinance == 0.2.49
markdownify == 0.13.1
pymongo == 4.10.1
ollama == 0.4.1
beautifulsoup4 ==4.12.3
requests == 2.32.3
python-dotenv == 1.0.1 
feedparser == 6.0.10

langchain-experimental == 0.3.4
langchain == 0.3.12
sentence_transformers == 3.3.1
gunicorn == 21.2.0
